chat_with_pdf/chat_with_pdf_updated/requirements.txt
langchain
flask
google-cloud-aiplatform
pypdf
PyPDF2
streamlit
faiss-cpu